Rename files add pytest
