+++ Dockerfile
@@ -12,5 +12,7 @@ COPY .pre-commit-config.yaml .
 RUN git init . && pre-commit install-hooks
 
 COPY worship_ppt/ /worship_ppt
+WORKDIR /worship_ppt
+RUN python3 -m pytest
 
 CMD gunicorn --bind :$PORT --capture-output --enable-stdio-inheritance --workers 1 --threads 8 "worship_ppt.app:create_app()";
