+++ requirements.txt
@@ -15,6 +15,7 @@ idna==2.10
 imageio==2.9.0
 imageio-ffmpeg==0.4.3
 importlab==0.7
+iniconfig==1.1.1
 isort==5.8.0
 itsdangerous==1.1.0
 Jinja2==2.11.3
@@ -32,15 +33,19 @@ ninja==1.10.2.3
 nodeenv==1.6.0
 numpy==1.20.1
 opencv-python==4.5.1.48
+packaging==21.3
 pandas==1.2.3
 Pillow==9.1.1
 platformdirs==2.5.2
+pluggy==1.0.0
 pre-commit==2.19.0
 proglog==0.1.9
+py==1.11.0
 pycodestyle==2.7.0
 pydot==1.4.2
 pylint==2.7.2
 pyparsing==3.0.9
+pytest==7.1.2
 python-dateutil==2.8.1
 python-docx==0.8.10
 python-pptx==0.6.18
@@ -51,6 +56,7 @@ requests==2.25.1
 six==1.15.0
 tabulate==0.8.10
 toml==0.10.2
+tomli==2.0.1
 tornado==6.1
 tqdm==4.59.0
 typing-extensions==3.10.0.0
